# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Our site
  version: 0.1.0
paths:
...
  /api/users/me/cart:
    get:
      tags:
      - Cart
      summary: Получить корзину текущего пользователя
      operationId: get_cart_me_api_users_me_cart_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
    delete:
      tags:
      - Cart
      summary: Очистить корзину текущего пользователя
      operationId: clear_cart_me_api_users_me_cart_delete
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
  /api/users/{user_id}/cart:
    get:
      tags:
      - Cart
      summary: Получить корзину пользователя по ID (Admin)
      operationId: get_cart_by_user_id_api_users__user_id__cart_get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
    delete:
      tags:
      - Cart
      summary: Очистить корзину пользователя по ID (Admin)
      operationId: clear_cart_by_user_id_api_users__user_id__cart_delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
  /api/users/me/cart/items:
    post:
      tags:
      - Cart
      summary: Добавить товар в корзину текущего пользователя
      operationId: add_item_to_cart_me_api_users_me_cart_items_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCartItemRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
  /api/users/{user_id}/cart/items:
    post:
      tags:
      - Cart
      summary: Добавить товар в корзину пользователя по ID (Admin)
      operationId: add_item_to_cart_by_user_id_api_users__user_id__cart_items_post
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCartItemRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
  /api/users/me/cart/items/{product_id}:
    patch:
      tags:
      - Cart
      summary: Обновить количество товара в корзине текущего пользователя
      operationId: update_cart_item_me_api_users_me_cart_items__product_id__patch
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Product Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
    delete:
      tags:
      - Cart
      summary: Удалить товар из корзины текущего пользователя
      operationId: remove_cart_item_me_api_users_me_cart_items__product_id__delete
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Product Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
  /api/users/{user_id}/cart/items/{product_id}:
    patch:
      tags:
      - Cart
      summary: Обновить количество товара в корзине пользователя по ID (Admin)
      operationId: update_cart_item_by_user_id_api_users__user_id__cart_items__product_id__patch
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Product Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
    delete:
      tags:
      - Cart
      summary: Удалить товар из корзины пользователя по ID (Admin)
      operationId: remove_cart_item_by_user_id_api_users__user_id__cart_items__product_id__delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Product Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
  /api/users/me/checkout:
    post:
      tags:
      - Checkout
      summary: Оформить заказ из товаров в корзине текущего пользователя
      operationId: checkout_cart_me_api_users_me_checkout_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
  /api/users/{user_id}/checkout:
    post:
      tags:
      - Checkout
      summary: Оформить заказ из товаров в корзине пользователя по ID (Admin)
      operationId: checkout_cart_by_user_id_api_users__user_id__checkout_post
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
...
components:
  schemas:
    AddCartItemRequest:
      properties:
        product_id:
          type: string
          format: uuid
          title: Product Id
          description: ID добавляемого товара
        quantity:
          type: integer
          exclusiveMinimum: 0.0
          title: Quantity
          description: Количество товара (должно быть > 0)
          default: 1
      type: object
      required:
      - product_id
      title: AddCartItemRequest
...
    CartItemResponse:
      properties:
        product_id:
          type: string
          format: uuid
          title: Product Id
        quantity:
          type: integer
          title: Quantity
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - product_id
      - quantity
      - updated_at
      title: CartItemResponse
    CartResponse:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        items:
          items:
            $ref: '#/components/schemas/CartItemResponse'
          type: array
          title: Items
        total_items:
          type: integer
          title: Total Items
      type: object
      required:
      - user_id
      - items
      - total_items
      title: CartResponse
...
    CheckoutResponse:
      properties:
        orders:
          items:
            $ref: '#/components/schemas/OrderResponse'
          type: array
          title: Orders
        total_orders:
          type: integer
          title: Total Orders
        total_amount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Total Amount
      type: object
      required:
      - orders
      - total_orders
      - total_amount
      title: CheckoutResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OrderResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        product_id:
          type: string
          format: uuid
          title: Product Id
        quantity:
          type: integer
          title: Quantity
        unit_price:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Unit Price
        total_amount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Total Amount
        status:
          $ref: '#/components/schemas/OrderStatus'
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - user_id
      - product_id
      - quantity
      - unit_price
      - total_amount
      - status
      - created_at
      title: OrderResponse
    OrderStatus:
      type: string
      enum:
      - CREATED
      - APPROVED
      - REJECTED
      - CANCELLED
      title: OrderStatus
...
    UpdateCartItemRequest:
      properties:
        quantity:
          type: integer
          minimum: 0.0
          title: Quantity
          description: Новое количество товара (если 0 — товар будет удален)
      type: object
      required:
      - quantity
      title: UpdateCartItemRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ErrorDetailResponse:
      title: ErrorDetailResponse
      type: object
      properties:
        detail:
          title: Detail
          type: string
          description: Описание ошибки
      required:
      - detail
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Access Token (в заголовке `Authorization: Bearer <token>`
        или cookie `access_token`)'